+++ src/hoodie/config.js
