+++ src/hoodie/connection.js
