+++ src/hoodie/dispose.js
