+++ test/specs/hoodie/dispose.spec.js
@@ -1,5 +1,5 @@
 require('../../lib/setup');
-var hoodieDispose = require('../../../src/utils/dispose');
+var hoodieDispose = require('../../../src/hoodie/dispose');
 
 describe('hoodie.dispose()', function() {
